# Add SSR performance measurement metrics (`Server-Timing` header)