--- a/packages/commerce-sdk-react/src/hooks/ShopperContexts/query.ts
+++ b/packages/commerce-sdk-react/src/hooks/ShopperContexts/query.ts
@@ -44,6 +44,11 @@ export const useShopperContext = (
     // We don't use `netOptions` here because we manipulate the options in `useQuery`.
     const method = async (options: Options) => await client[methodName](options)
 
+    queryOptions.meta = {
+        displayName: 'useShopperContext',
+        ...queryOptions.meta
+    }
+
     // For some reason, if we don't explicitly set these generic parameters, the inferred type for
     // `Data` sometimes, but not always, includes `Response`, which is incorrect. I don't know why.
     return useQuery<Client, Options, Data>({...netOptions, parameters}, queryOptions, {