--- a/packages/pwa-kit-react-sdk/setup-jest.js
+++ b/packages/pwa-kit-react-sdk/setup-jest.js
@@ -8,6 +8,7 @@
 
 import 'regenerator-runtime/runtime'
 import '@testing-library/jest-dom'
+import {performance} from 'perf_hooks'
 // Mock the application configuration to be used in all tests.
 jest.mock('@salesforce/pwa-kit-runtime/utils/ssr-config', () => {
     return {
@@ -39,3 +40,8 @@ jest.mock('@salesforce/pwa-kit-runtime/utils/ssr-config', () => {
         })
     }
 })
+
+// The global performance object is available in production
+// environments for both the server and the client.
+// It's just the jest environment that this is not available
+global.performance = performance