… packages/commerce-sdk-react/CHANGELOG.md
… packages/commerce-sdk-react/src/hooks/ShopperBaskets/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperContexts/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperCustomers/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperExperience/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperGiftCertificates/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperLogin/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperOrders/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperProducts/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperPromotions/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperSearch/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperSeo/query.test.ts
… packages/commerce-sdk-react/src/hooks/ShopperStores/query.test.ts
… packages/pwa-kit-react-sdk/CHANGELOG.md
… packages/pwa-kit-react-sdk/src/ssr/server/react-rendering.test.js
… packages/pwa-kit-react-sdk/src/ssr/universal/components/with-legacy-get-props/index.js
… packages/pwa-kit-react-sdk/src/ssr/universal/components/with-react-query/index.js
… packages/pwa-kit-react-sdk/src/utils/performance.js
… packages/pwa-kit-react-sdk/src/utils/performance.test.js