# FIRST, PLEASE READ https://github.com/SalesforceCommerceCloud/pwa-kit/pull/1885 

This is another approach for introducing ssr performance metrics, thanks to the feedback @johnboxall gave!

The performance tracking is controlled by environment variable `SERVER_TIMING` or a query parameter `__server_timing`

This PR adds a conditional [Server-Timing HTTP header](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Server-Timing). This header is included when a query parameter is used `?__server_timing`.

#### http response header
![CleanShot 2024-07-16 at 12 32 50](https://github.com/user-attachments/assets/64a40f2a-6fdf-4677-ac40-0e0c459eca37)


#### console logs

```
pwa-kit-react-sdk.performance DEBUG ssr:route-matching - 0.15ms
pwa-kit-react-sdk.performance DEBUG ssr:load-component - 0.09ms
pwa-kit-react-sdk.performance DEBUG ssr:fetch-strategies:react-query:pre-render - 99.14ms
pwa-kit-react-sdk.performance DEBUG ssr:fetch-strategies:get-prop::App - 0.01ms
pwa-kit-react-sdk.performance DEBUG ssr:fetch-strategies:get-prop::home - 1000.70ms
pwa-kit-react-sdk.performance DEBUG ssr:fetch-strategies:react-query:use-query::1 - 13.98ms ["app","translations","messages","en-GB"]
pwa-kit-react-sdk.performance DEBUG ssr:fetch-strategies:react-query:use-query::useCategory - 667.09ms ["/commerce-sdk-react","/organizations/","f_ecom_zzrf_001","/categories/","root",{"id":"root","levels":1,"locale":"en-GB","organizationId":"f_ecom_zzrf_001","siteId":"RefArchGlobal"}]
pwa-kit-react-sdk.performance DEBUG ssr:fetch-strategies:react-query:use-query::useProductSearch - 1180.25ms ["/commerce-sdk-react","/organizations/","f_ecom_zzrf_001","/product-search",{"allImages":true,"allVariationProperties":true,"currency":"GBP","expand":["promotions","variations","prices","images","custom_properties"],"limit":10,"locale":"en-GB","organizationId":"f_ecom_zzrf_001","perPricebook":true,"refine":["cgid=newarrivals","htype=master"],"siteId":"RefArchGlobal"}]
pwa-kit-react-sdk.performance DEBUG ssr:fetch-strategies - 1282.28ms
pwa-kit-react-sdk.performance DEBUG ssr:render-to-string - 120.91ms
pwa-kit-react-sdk.performance DEBUG ssr:total - 1444.56ms
```


- for `getProps`, the page name and the `App` component will show up as a part of the metrics name
- for `useQuery`, the query cache key is added to the logs.


## Steps to reproduce

### Query parameter tests

1. run `npm start` in template-retail-react-app and template-typescript-minimal
2. open localhost:3000?__server_timing
3. verify the `Server-Timing` http response header is populated
4. verify the console contain relevant logs

### Env var tests
1. go to ./packages/template-retail-react-app
2. update npm start commands to `"start": "cross-env SERVER_TIMING=1 pwa-kit-dev start",`
4. verify the console contain relevant logs